#166775120: implement classes and put all the logics in the same file [Finishes #166775120]
