+++ server/models/tables.js
@@ -21,8 +21,7 @@ const createTables = `
       price INT NOT NULL,
       manufacturer VARCHAR(20) NOT NULL,
       model VARCHAR(20) NOT NULL,
-      body_type VARCHAR(20) NOT NULL,
-      FOREIGN KEY (owner) REFERENCES users(id) ON DELETE CASCADE
+      body_type VARCHAR(20) NOT NULL
     );
 
     CREATE TABLE IF NOT EXISTS orders(
@@ -30,8 +29,7 @@ const createTables = `
       buyer INT NOT NULL,
       car_id INT NOT NULL,
       amount INT NOT NULL,
-      status VARCHAR(20) NOT NULL,
-      FOREIGN KEY (buyer) REFERENCES users(id) ON DELETE CASCADE
+      status VARCHAR(20) NOT NULL
     );
 
     CREATE TABLE IF NOT EXISTS flags(
